+++ OtterEngine/Entity/Triangle.cpp
@@ -33,7 +33,7 @@ Triangle::Triangle(const Graphics& graphics, const Vector3& translation, const V
 
 		// buffers
 		s_commonResources.push_back(std::make_unique<VertexBuffer>(graphics, 
-			s_vertices.data(), sizeof(Vertex), s_vertices.size()));
+			s_vertices.data(), static_cast<unsigned int>(sizeof(Vertex)), s_vertices.size()));
 		s_commonResources.push_back(std::make_unique<IndexBuffer>(graphics, s_indices));
 	}
